refactor : enums 패키지 만들어 그 안으로 이동
