+++ src/main/java/sisosolsol/greenfire/common/enums/CategoryType.java
@@ -1,4 +1,4 @@
-package sisosolsol.greenfire.category.model.type;
+package sisosolsol.greenfire.common.enums;
 
 public enum CategoryType {
     CHALLENGE,  // 챌린지
